Support for aggregating results from multijob downstream builds
Multijobs can be triggered from the flow job and we would like to collect the results from all the jobs.

When the multijob is invoked from the flow, results are collected from 2 places:
* multijob parent - extra results can be present (ideally should not clone the results stored at downstream builds)
* multijob subbuilds

Another recursion would be possible as well (you can trigger another multijob or flow from the multijob), but this part is not resolved at this change.